# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  CreateServerRequest:
    properties:
      authToken:
        type: string
      baseURL:
        type: string
      headersJSON:
        type: string
      name:
        type: string
      status:
        type: string
    required:
    - baseURL
    - name
    type: object
...
  DeleteServerResponse:
    properties:
      deleted:
        type: boolean
    required:
    - deleted
    type: object
  DeleteServerResponseDoc:
    properties:
      data:
        $ref: '#/definitions/DeleteServerResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  McpErrorDoc:
    properties:
      errorMsg:
        type: string
    required:
    - errorMsg
    type: object
...
  ServerDataResponse:
    properties:
      server:
        $ref: '#/definitions/ServerResponse'
    required:
    - server
    type: object
  ServerDataResponseDoc:
    properties:
      data:
        $ref: '#/definitions/ServerDataResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  ServerListResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/ServerResponse'
        type: array
    required:
    - results
    type: object
  ServerListResponseDoc:
    properties:
      data:
        $ref: '#/definitions/ServerListResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  ServerResponse:
    properties:
      activeToolCount:
        type: integer
      baseURL:
        type: string
      createdAt:
        type: string
      headersJSON:
        type: string
      id:
        type: integer
      lastError:
        type: string
      lastSyncedAt:
        type: string
        x-nullable: true
        x-omitempty: false
      name:
        type: string
      requiresToolMetadataSyncConfirmation:
        type: boolean
      sortOrder:
        type: integer
      status:
        type: string
      toolCount:
        type: integer
      updatedAt:
        type: string
    required:
    - activeToolCount
    - baseURL
    - createdAt
    - headersJSON
    - id
    - lastError
    - lastSyncedAt
    - name
    - requiresToolMetadataSyncConfirmation
    - sortOrder
    - status
    - toolCount
    - updatedAt
    type: object
...
info:
  contact: {}
  description: DEEIX Chat 后端 API 文档
  title: DEEIX Chat API
  version: "0.4.0"
paths:
...
  /admin/mcp/servers:
    get:
      description: 管理员查看已配置的 MCP 服务及其工具统计
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ServerListResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/McpErrorDoc'
      security:
      - BearerAuth: []
      summary: 获取 MCP 服务列表
      tags:
      - admin-mcp
    post:
      consumes:
      - application/json
      description: 管理员创建一个 MCP 服务配置
      parameters:
      - description: MCP 服务配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ServerDataResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/McpErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/McpErrorDoc'
      security:
      - BearerAuth: []
      summary: 创建 MCP 服务
      tags:
      - admin-mcp
  /admin/mcp/servers/{id}:
    delete:
      description: 管理员删除一个 MCP 服务及其工具
      parameters:
      - description: MCP 服务 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DeleteServerResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/McpErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/McpErrorDoc'
      security:
      - BearerAuth: []
      summary: 删除 MCP 服务
      tags:
      - admin-mcp
    patch:
      consumes:
      - application/json
      description: 管理员更新一个 MCP 服务配置
      parameters:
      - description: MCP 服务 ID
        in: path
        name: id
        required: true
        type: integer
      - description: MCP 服务配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ServerDataResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/McpErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/McpErrorDoc'
      security:
      - BearerAuth: []
      summary: 更新 MCP 服务
      tags:
      - admin-mcp
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"